Will pursue developement with my own lilbrary, mixed with CL library, at least in terms of program and kernel management, which will allow me to use #defines and other preprocessor macros in my CL files (I think).

+++ enjacl/graphics_libs/opencl_utilities/cl.h
@@ -87,7 +87,7 @@ public:
 	//void profile();
 	void profile(cl_event event= (cl_event)0);
 	void printDeviceID() {
-		printf("deviceID: %d\n", (int) device_id);
+		printf("deviceID: %d\n", (long) device_id);
 	}
 	//void printKernel(const char* msg="") {
 		//printf("%s, kernel: %d\n", msg, (int) kernel);
